+++ src/digital/mod.rs
@@ -3,14 +3,25 @@
 //!
 //!
 
-// Fallible digital traits
-// This has been left as a submodule to smooth transitions from v0.2.x and
-// may be removed in future
+// Deprecated / infallible traits
 #[deprecated(
-    since = "0.3.0",
-    note = "Please use traits directly from `digital::` instead"
+    since = "0.2.2",
+    note = "Deprecated because the methods cannot return errors. \
+            Users should use the traits in digital::v2."
 )]
+pub mod v1;
+
+// New / fallible traits
 pub mod v2;
 
-// Re-export default traits
+// v2 -> v1 compatibility wrappers
+// These require explicit casts from v2 -> v1
+pub mod v1_compat;
+
+// v1 -> v2 compatibility shims
+// These are implicit over v1 implementations
+pub mod v2_compat;
+
+// Re-export old traits so this isn't a breaking change
+#[allow(deprecated)]
 pub use self::v2::*;